// src/app/intefaces/venta.interface.ts
export class Venta {
    key: string;
    keyCliente: string;
    fecha: number;
    pagado: Boolean;
    tienda: string;
    vendedor: string;
    total: number;
    estado: string;
    calleNum: string;
    numeroInterior?: string;
    ciudad: number;
    cp: number;
    estadoCliente: string;
    telefono: string;
    nombreCliente: string;
    keyPublico: string;
    fechaPago: number;
    fechaEnvio: number;
    fechaEntrega: number;

    tipoEnvio: string;
    costoEnvio: number;
    numeroGuia: string;
    paqueteriaEnvio: string;


}